+++ jest.config.js
@@ -17,7 +17,7 @@ module.exports = {
   },
   coverageThreshold: {
     global: {
-      branches: 75,
+      branches: 50,
       functions: 80,
       lines: 80,
       statements: 80,