+++ src/contexts/inspection/inspection.tsx
@@ -1,14 +1,15 @@
 import { createContext, Dispatch } from "react";
 
+import axiosClient from "~/config/axios";
 import { InspectionStateType } from "~/contexts/inspection/types";
 import { InspectionAction } from "~/reducers/inspector/types";
-import InspectionService from "~/services/inspection-service/inspection-service";
+import InspectionService from "~/services/inspection/inspection";
 
 export const initialState: InspectionStateType = {
   isLoading: false,
   error: false,
   inspections: [],
-  inspectionService: new InspectionService(),
+  inspectionService: new InspectionService(axiosClient),
   inspectionDispatch: (() => undefined) as Dispatch<InspectionAction>,
 };