+++ src/contexts/inspection/types.ts
@@ -2,7 +2,7 @@ import { Dispatch } from "react";
 
 import { Inspection } from "~/entities/inspection";
 import { InspectionAction } from "~/reducers/inspector/types";
-import InspectionService from "~/services/inspection-service/inspection-service";
+import InspectionService from "~/services/inspection/inspection";
 
 export type InspectionStateType = {
   isLoading: boolean;